+++ src/cjs/Bounds.d.ts
@@ -10,6 +10,7 @@
  * @modified 2022-11-28 Added the `clone` method.
  * @modified 2023-09-29 Added the `randomPoint` method.
  * @modified 2025-03-23 Added the `getMinDimension` and `getMaxDimension` methods.
+ * @modified 2025-04-18 Change parameter type in `Bounds.computeFromVertices` from `Vertex` to more general `XYCoords`.
  * @version  1.8.0
  **/
 import { Polygon } from "./Polygon";
@@ -131,10 +132,10 @@ export declare class Bounds implements IBounds, XYDimension {
      * @static
      * @method computeFromVertices
      * @memberof Bounds
-     * @param {Array<Vertex>} vertices - The set of vertices you want to get the bounding box for.
+     * @param {Array<XYCoords>} vertices - The set of vertices you want to get the bounding box for.
      * @return The minimal Bounds for the given vertices.
      **/
-    static computeFromVertices(vertices: Array<Vertex>): Bounds;
+    static computeFromVertices(vertices: Array<XYCoords>): Bounds;
     /**
      * Create a new `Bounds` instance just from `width` and `height`, located at (0,0) or the optionally given origin.
      *